# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "rest-API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/teams": {
            "get": {
                "description": "Получить команды, в которых состоит текущий пользователь",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Get user's teams",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/mkk_basis_rest_api_internal_common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/mkk_basis_rest_api_internal_app_core_entities_teams-entities.TeamResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Создать команду; текущий пользователь становится owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Create team",
                "parameters": [
                    {
                        "description": "Team data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mkk_basis_rest_api_internal_app_core_entities_teams-entities.CreateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/mkk_basis_rest_api_internal_common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/mkk_basis_rest_api_internal_app_core_entities_teams-entities.TeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/teams/{id}/invite": {
            "post": {
                "description": "Добавить пользователя в команду; доступно owner и admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Invite user to team",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invitation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mkk_basis_rest_api_internal_app_core_entities_teams-entities.InviteUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/mkk_basis_rest_api_internal_common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/mkk_basis_rest_api_internal_app_core_entities_teams-entities.TeamMemberResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mkk_basis_rest_api_internal_app_core_entities_teams-entities.CreateTeamRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "mkk_basis_rest_api_internal_app_core_entities_teams-entities.InviteUserRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "member"
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "mkk_basis_rest_api_internal_app_core_entities_teams-entities.TeamMemberResponse": {
            "type": "object",
            "properties": {
                "joined_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "mkk_basis_rest_api_internal_app_core_entities_teams-entities.TeamResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "joined_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "mkk_basis_rest_api_internal_common.APIResponse": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "result": {}
            }
        }
    }
}